pyproject.toml
#
# This file is part of pleiades_aligner
# by Tom Elliott for the Institute for the Study of the Ancient World
# (c) Copyright 2023 by New York University
# Licensed under the AGPL-3.0; see LICENSE.txt file.
#

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pleiades_aligner"
version = "0.0.1"
authors = [
  { name="Tom Elliott", email="tom.elliott@nyu.edu" },
]
description = "Software tools to align external datasets with the Pleiades gazetteer"
readme = "README.md"
requires-python = ">=3.12.0"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Operating System :: OS Independent",
]
dependencies = [
  "airtight",
  "colorama",
  "encoded_csv @ https://github.com/paregorios/encoded_csv/archive/refs/heads/main.zip",
  "haversine",
  #"iteration_utilities",
  #"fiona",
  #"pleiades_entities @ file:///Users/paregorios/Documents/files/P/pleiades_entities",
  "pleiades_local @ file:///Users/paregorios/Documents/files/P/pleiades_local",
  "pytest",
  #"reproj",
  "shapely",
  "platformdirs",
  "python-slugify",
  "textnorm",
  #"validators",
  ##"webiquette @ https://github.com/isawnyu/webiquette/archive/refs/heads/main.zip"
  #"webiquette @ file:///Users/paregorios/Documents/files/W/webiquette"
]
[project.urls]
# "Homepage" = "https://github.com/pypa/sampleproject"
# "Bug Tracker" = "https://github.com/pypa/sampleproject/issues"
[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"